refactor(frontend): split artifact detail into reusable UI components
